Time: 32 ms (38.13%), Space: 47.9 MB (5.12%) - LeetHub

+++ 1408-find-the-smallest-divisor-given-a-threshold/1408-find-the-smallest-divisor-given-a-threshold.java
@@ -9,7 +9,7 @@ class Solution {
     public static boolean isPossible(int[] nums,int divisor,int threshold){
         long sum = 0;
         for (int i = 0; i < nums.length; i++) {
-            sum += Math.ceil((double) nums[i] / divisor);
+            sum += Math.ceil((double) nums[i] / (double)divisor);
         }
         return sum <= threshold;
     }
